refactor: 카드 등록 페이지 context provider감싸도록 컴포넌트 분리
